requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==4.4.2
Django==4.1.3
djangorestframework==3.14.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.4
Pillow==9.3.0
proglog==0.1.10
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.12
